Update pagination.php
By setting $config['attributes'] like this, every anchor in the pagination will have the 'page-link' class added to itself. You only have to use the "<span class="page-link">" method on the "cur_tag_open" since the content is not a link.

The reason this way is better is due to the fact that the previous method rendered only the links clickable, not the whole 'page-item' button, besides it being less of a workaround. Using the previous method on the cur_tag is not a problem because none of it should be clickable anyway.